# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "jobtracker-api",
    "version": "0.1.0"
  },
  "paths": {
    "/v1/users/bootstrap": {
      "post": {
        "summary": "Bootstrap",
        "operationId": "bootstrap_v1_users_bootstrap_post",
        "parameters": [
          {
            "name": "x-service-token",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-Service-Token"
            }
          },
          {
            "name": "x-user-sub",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-User-Sub"
            }
          },
          {
            "name": "x-user-email",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-User-Email"
            }
          },
          {
            "name": "x-user-name",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-User-Name"
            }
          },
          {
            "name": "x-user-groups",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-User-Groups"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/user/usage": {
      "get": {
        "summary": "Usage",
        "operationId": "usage_v1_user_usage_get",
        "parameters": [
          {
            "name": "x-service-token",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-Service-Token"
            }
          },
          {
            "name": "x-user-sub",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-User-Sub"
            }
          },
          {
            "name": "x-user-email",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-User-Email"
            }
          },
          {
            "name": "x-user-name",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-User-Name"
            }
          },
          {
            "name": "x-user-groups",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-User-Groups"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/models": {
      "get": {
        "summary": "Models",
        "description": "Only the options valid for this user right now: their BYO provider's\ncatalog if they have a key (it takes precedence), else the owner-key\nallowlist if granted, else nothing runnable.",
        "operationId": "models_v1_models_get",
        "parameters": [
          {
            "name": "x-service-token",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-Service-Token"
            }
          },
          {
            "name": "x-user-sub",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-User-Sub"
            }
          },
          {
            "name": "x-user-email",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-User-Email"
            }
          },
          {
            "name": "x-user-name",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-User-Name"
            }
          },
          {
            "name": "x-user-groups",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-User-Groups"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/user/settings": {
      "get": {
        "summary": "Get Settings",
        "operationId": "get_settings_v1_user_settings_get",
        "parameters": [
          {
            "name": "x-service-token",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-Service-Token"
            }
          },
          {
            "name": "x-user-sub",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-User-Sub"
            }
          },
          {
            "name": "x-user-email",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-User-Email"
            }
          },
          {
            "name": "x-user-name",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-User-Name"
            }
          },
          {
            "name": "x-user-groups",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-User-Groups"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Put Settings",
        "operationId": "put_settings_v1_user_settings_put",
        "parameters": [
          {
            "name": "x-service-token",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-Service-Token"
            }
          },
          {
            "name": "x-user-sub",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-User-Sub"
            }
          },
          {
            "name": "x-user-email",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-User-Email"
            }
          },
          {
            "name": "x-user-name",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-User-Name"
            }
          },
          {
            "name": "x-user-groups",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-User-Groups"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SettingsPut"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/user/settings/api-key": {
      "put": {
        "summary": "Put Api Key",
        "operationId": "put_api_key_v1_user_settings_api_key_put",
        "parameters": [
          {
            "name": "x-service-token",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-Service-Token"
            }
          },
          {
            "name": "x-user-sub",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-User-Sub"
            }
          },
          {
            "name": "x-user-email",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-User-Email"
            }
          },
          {
            "name": "x-user-name",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-User-Name"
            }
          },
          {
            "name": "x-user-groups",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-User-Groups"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApiKeyPut"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete Api Key",
        "operationId": "delete_api_key_v1_user_settings_api_key_delete",
        "parameters": [
          {
            "name": "x-service-token",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-Service-Token"
            }
          },
          {
            "name": "x-user-sub",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-User-Sub"
            }
          },
          {
            "name": "x-user-email",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-User-Email"
            }
          },
          {
            "name": "x-user-name",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-User-Name"
            }
          },
          {
            "name": "x-user-groups",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-User-Groups"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApiKeyPut": {
        "properties": {
          "api_key": {
            "type": "string",
            "maxLength": 400,
            "minLength": 8,
            "title": "Api Key"
          },
          "provider": {
            "type": "string",
            "title": "Provider",
            "default": "openai"
          },
          "base_url": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 400
              },
              {
                "type": "null"
              }
            ],
            "title": "Base Url"
          }
        },
        "type": "object",
        "required": [
          "api_key"
        ],
        "title": "ApiKeyPut"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ModelEntry": {
        "properties": {
          "model": {
            "type": "string",
            "title": "Model"
          },
          "note": {
            "type": "string",
            "title": "Note"
          }
        },
        "type": "object",
        "required": [
          "model",
          "note"
        ],
        "title": "ModelEntry"
      },
      "ModelsResponse": {
        "properties": {
          "providers": {
            "items": {
              "$ref": "#/components/schemas/ProviderEntry"
            },
            "type": "array",
            "title": "Providers"
          },
          "owner_key_models": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Owner Key Models"
          },
          "key_source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Key Source"
          },
          "addable_providers": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Addable Providers"
          }
        },
        "type": "object",
        "required": [
          "providers",
          "owner_key_models",
          "key_source",
          "addable_providers"
        ],
        "title": "ModelsResponse"
      },
...
      "ProviderEntry": {
        "properties": {
          "provider": {
            "type": "string",
            "title": "Provider"
          },
          "default_model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Default Model"
          },
          "models": {
            "items": {
              "$ref": "#/components/schemas/ModelEntry"
            },
            "type": "array",
            "title": "Models"
          },
          "params": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Params"
          }
        },
        "type": "object",
        "required": [
          "provider",
          "default_model",
          "models",
          "params"
        ],
        "title": "ProviderEntry"
      },
...
      "SettingsPut": {
        "properties": {
          "column_layout": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Column Layout"
          },
          "prefs": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prefs"
          },
          "ai_model": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200
              },
              {
                "type": "null"
              }
            ],
            "title": "Ai Model"
          },
          "ai_params": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ai Params"
          },
          "bypass_sponsorship_filter": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bypass Sponsorship Filter"
          }
        },
        "type": "object",
        "title": "SettingsPut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}